pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-airflow>=3.2.1",
    "faker>=40.15.0",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.12",
    "sqlalchemy>=2.0.49",
]
